+++ oasis-types/src/lib.rs
@@ -30,50 +30,3 @@ pub struct Event {
     pub topics: Vec<[u8; 32]>,
     pub data: Vec<u8>,
 }
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-
-    #[test]
-    fn deserialize_address_from_array() {
-        let bytes = [1u8; 20];
-        let addr: Address = serde_cbor::from_slice(&serde_cbor::to_vec(&bytes).unwrap()).unwrap();
-        assert_eq!(addr.0, bytes);
-    }
-
-    #[test]
-    #[should_panic]
-    fn fail_deserialize_address_from_short_array() {
-        let bytes = [1u8; 19];
-        serde_cbor::from_slice::<Address>(&serde_cbor::to_vec(&bytes).unwrap()).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn fail_deserialize_address_from_long_array() {
-        let bytes = [1u8; 21];
-        serde_cbor::from_slice::<Address>(&serde_cbor::to_vec(&bytes).unwrap()).unwrap();
-    }
-
-    #[test]
-    fn deserialize_address_from_slice() {
-        let bytes = vec![1u8; 20];
-        let addr: Address = serde_cbor::from_slice(&serde_cbor::to_vec(&bytes).unwrap()).unwrap();
-        assert_eq!(&addr.0, bytes.as_slice());
-    }
-
-    #[test]
-    #[should_panic]
-    fn fail_deserialize_address_from_short_slice() {
-        let bytes = vec![1u8; 19];
-        serde_cbor::from_slice::<Address>(&serde_cbor::to_vec(&bytes).unwrap()).unwrap();
-    }
-
-    #[test]
-    #[should_panic]
-    fn fail_deserialize_address_from_long_slice() {
-        let bytes = vec![1u8; 21];
-        serde_cbor::from_slice::<Address>(&serde_cbor::to_vec(&bytes).unwrap()).unwrap();
-    }
-}
